improve scope identification
This is a big improvement for trying to find refs inside another ref
(e.g. `referral --scope Haystack#hide` will now give everything inside
that method definition).

However there is a downside which the "parent" node of some things will
appear illogical. For example, `Hay = Struct.new()` will yield: 
`call Hay Struct.new`, as if the Struct.new call's parent is `Hay`.
From the AST's perspective this is right, but it seems off if you only
think of scope in terms of "inside of"